Services
Created a service class that can be shared by many components.
Used the ngOnInit lifecycle hook to get the hero data when the AppComponent activates.
Defined the HeroService as a provider for the AppComponent.
Created mock hero data and imported them into the service.
Designed the service to return a Promise and the component to get the data from the Promise.